requirements.txt
gitpython
colorama
